Fix lint

+++ src/api/deposits.ts
@@ -56,7 +56,7 @@ export const useDeposits = (poolId: u32 | string) => {
   const api = useApiPromise()
   return useQuery(QUERY_KEYS.deposits(poolId), getDeposits(api, poolId))
 }
-export const useOmniPositionId = (positionId: u32 | string) => {
+export const useOmniPositionId = (positionId: u128 | string) => {
   const api = useApiPromise()
 
   return useQuery(
@@ -76,7 +76,7 @@ const getDeposits = (api: ApiPromise, poolId: u32 | string) => async () => {
 }
 
 const getOmniPositionId =
-  (api: ApiPromise, poolId: u32 | string) => async () => {
+  (api: ApiPromise, poolId: u128 | string) => async () => {
     const res = await api.query.omnipoolLiquidityMining.omniPositionId(poolId)
     return res
   }
